+++ cypress/integration/parallax.spec.js
@@ -29,18 +29,22 @@ describe("Parallax", () => {
   });
 
   it("should call onProgress with expected values", () => {
-    cy.get('.progress-details').should('contain', '0 1 0');
+    cy.get('.parallax-progress-details').should('contain', '0 1 0');
+    cy.get('.layer-progress-details').should('contain', '0');
 
     cy.scrollTo(0, HEIGHT / 2);
 
-    cy.get('.progress-details').should('contain', '0.25 1 0.5');
+    cy.get('.parallax-progress-details').should('contain', '0.25 1 0.5');
+    cy.get('.layer-progress-details').should('contain', '0');
 
     cy.scrollTo(0, HEIGHT);
 
-    cy.get('.progress-details').should('contain', '0.5 2 0');
+    cy.get('.parallax-progress-details').should('contain', '0.5 2 0');
+    cy.get('.layer-progress-details').should('contain', '0.5');
 
     cy.scrollTo('bottom');
 
-    cy.get('.progress-details').should('contain', '1 3 0');
+    cy.get('.parallax-progress-details').should('contain', '1 3 0');
+    cy.get('.layer-progress-details').should('contain', '1');
   });
 });