+++ src/utils/clamp.js
@@ -1,0 +1,3 @@
+export const clamp = (val, min, max) => {
+  return val < min ? min : val > max ? max : val;
+};